bookreader updates

+++ createxml.php
@@ -5,25 +5,19 @@ set_time_limit(0);
 ini_set("memory_limit","1024M");
 date_default_timezone_set('US/Central');
 
-$linefeed = PHP_EOL;
-$space = " ";
+$linefeed = "<br />";
+$space = "&nbsp;";
 
 echo "Begin:" . $linefeed;
 
 $counter = 0;
 $totalcount = 0;
 
-$startNum = 99;  //71 = 1962
-$endCount = 10;  // Number of yearbooks to process
+$startNum = 0;  
+$endCount = 1;  // Number of directories to process
 
-//$dirf    			= 'c:/zz/';
-//$genericimageloc    = 'c:/zz/';
-//$tmpfolder 			= 'c:/tmp/';
-
-$dirf    			= DIRF;
-$genericimageloc    = '/tmp/';
-$tmpfolder 			= '/tmp/';
-
+// this is defined in the config.php
+$dirf    = DIRF;
 $dir = scandir($dirf);
 
 	foreach($dir as $file) {
@@ -31,34 +25,31 @@ $dir = scandir($dirf);
 
 		//echo "Total is: " . $totalcount . " and Start Number is: " . $startNum . $linefeed;		
 		
-			if (strrpos( $file, 'yb') === 0) {
+			if (strrpos( $file, DIRECTORY_IDENTIFIER) === 0) {
 				
 				if ( $totalcount >= $startNum ) {
 					echo "Pass Gate:" . $linefeed;	
 					
-					mkdir($tmpfolder . $file, 0700);
-				
-					echo "$space$spaceProcessing Yearbook:  " . $file . $linefeed;
+					echo $space . $space . "Processing Yearbook:  " . $file . $linefeed;
 					$dir2 = scandir($dirf . "/" . $file);
-					echo "$space$space$spaceDeleting Previous Files:" . $linefeed;
+					echo $space . $space . $space . $space . "Deleting Previous Files:" . $linefeed;
+					
 						foreach($dir2 as $file2) {
-							if ( ($file2 != $file . "_abbyy.gz") && ($file2 != $file . "_abbyy.xml") && ($file2 != ".") && ($file2 != "..") && ($file2 != $file . "_jp2.zip") && ($file2 != $file . "_raw_jp2.zip") ) {
-
-								echo "$space$space$space$spaceFile is:  " . $file2 . $linefeed;
-
-								if (file_exists($dirf . "/" . $file . "/" . $file2)) {
-									unlink($dirf . "/" . $file . "/" . $file2);
-									echo "$space$space$space$space$space$file2$spaceDeleted" . $linefeed;
-								}					
+							if ( ($file2 != $file . ".php") && ($file2 != $file . "_abbyy.gz") && ($file2 != $file . "_abbyy.xml") && ($file2 != ".") && ($file2 != "..") && ($file2 != $file . "_jp2.zip") && ($file2 != $file . "_raw_jp2.zip") ) {
+								echo $space . $space . $space . $space . $space . "File is:  " . $file2 . $linefeed;
+									if (file_exists($dirf . "/" . $file . "/" . $file2)) {
+										unlink($dirf . "/" . $file . "/" . $file2);
+										echo $space . $space . $space . $space . $space . $file2 . $space . "Deleted" . $linefeed;
+									}
 							}
 						}
 					
-					echo "$linefeed$space$spaceLooking for zip file: " . $dirf . $file . '/' . $file . "_jp2.zip" . $linefeed;
+					echo $linefeed . $space . $space . "Looking for zip file: " . $dirf . '/' . $file . '/' . $file . "_jp2.zip" . $linefeed;
 					
 					$zip = new ZipArchive;
 					$res = $zip->open($dirf . "/" . $file . "/" . $file . "_jp2.zip");
 						if ($res === TRUE) {
-							echo "$space$space$space$spaceZip is Open with "; 
+							echo $space . $space . $space . $space . "Zip is Open with "; 
 							echo $zip->numFiles . " files" . $linefeed;
 							$counter = $zip->numFiles;
 
@@ -66,43 +57,39 @@ $dir = scandir($dirf);
 								
 							$im_string = $zip->getFromName($file . '_jp2/' . $xFile);
 						
-							$ofp = fopen( $tmpfolder . $file . '/' . $xFile, 'w' );
-							//$ofp = fopen( $dirf . $file . '/' . $xFile, 'w' );
+							$ofp = fopen( $dirf . "/" . $file . '/' . $xFile, 'w' );
 
 							fwrite( $ofp, $im_string );
 							fclose($ofp); 
-							echo "$space$spaceCover Image extracted" . $linefeed;
+							echo $space . $space . "Cover Image extracted" . $linefeed;
 							$zip->close();
 						} else {
-							echo 'failed cover image with code:' . $res . $linefeed;
+							echo $space . $space . "failed cover image with code:" . $res . $linefeed;
 						}
 						
 					
 
-					echo "$space$spaceCreate XML" . $linefeed;
-					$xml = CreateXML($dirf . $file . "/" . $file . "_abbyy.xml", $file, $counter);
-			//		file_put_contents($dirf . $file . '/scandata.xml', $xml);
-					file_put_contents($tmpfolder . $file . '/scandata.xml', $xml);
+					echo $space . $space . "Create XML" . $linefeed;
+					$xml = CreateXML($dirf . "/" . $file . "/" . $file . "_abbyy.xml", $file, $counter);
+					file_put_contents($dirf . "/" . $file . '/scandata.xml', $xml);
 					
-					echo "$space$spaceCreate CreateFilesXML" . $linefeed;
+					echo $space . $space . "Create FilesXML" . $linefeed;
 					$xml = CreateFilesXML($file, $dirf);
-			//		file_put_contents($dirf . $file . '/' . $file . '_files.xml', $xml);
-					file_put_contents($tmpfolder . $file . '/' . $file . '_files.xml', $xml);
+					file_put_contents($dirf . "/" . $file . '/' . $file . '_files.xml', $xml);
 					
-					echo "$space$spaceCreate CreateMetadataFilesXML" . $linefeed;
+					echo $space . $space . "Create MetadataFilesXML" . $linefeed;
 					$xml = CreateMetadataFilesXML($file, $SQLIPAddress, $UserName, $Password, $db);
-			//		file_put_contents($dirf . $file . '/' . $file . '_meta.xml', $xml);			
-					file_put_contents($tmpfolder . $file . '/' . $file . '_meta.xml', $xml);	
+					file_put_contents($dirf . "/" . $file . '/' . $file . '_meta.xml', $xml);			
 					
-					echo "$space$spaceCreate CreateImageCMD" . $linefeed;
+					echo $space . $space . "Create ImageCMD" . $linefeed;
 					CreateImageCMD($file, $dirf, $tmpfolder);
 					
 //					CreateImage($file, $dirf);
 			
-					echo "$space$spaceCreate CreateZip" . $linefeed;
+					echo $space . $space . "Create Zip" . $linefeed;
 					CreateZip($file, $dirf, $genericimageloc);
 					
-					echo "$space$spaceCreate CreateAbbyygz" . $linefeed;
+					echo $space . $space . "Create Abbyy.gz" . $linefeed;
 					CreateAbbyygz($file, $dirf, $tmpfolder);
 					
 					echo $linefeed;
@@ -112,18 +99,21 @@ $dir = scandir($dirf);
 				//echo $linefeed;
 					
 			} else {
-				$totalcount = $totalcount + 1;
-			//	echo "$space$space$space$space$spaceSkipping Directory" . $linefeed;
+				
+			//	$totalcount = $totalcount + 1;
+				echo $space . $space . $space . $space . $space . "Skipping Directory" . $linefeed;
 			}
 		
 		$counter = 0;
 		
 		if ( ($startNum + $endCount) <= $totalcount ) {
-			echo ($startNum + $endCount) . $linefeed;
-			echo  $totalcount . $linefeed;
-			echo "break";
+			echo $linefeed;
+			//echo ($startNum + $endCount) . $linefeed;
+			//echo  $totalcount . $linefeed;
+			echo "I have done " . ($startNum + $endCount) . " and I should break" . $linefeed;
 			break;
 		}
+		
 	}
 
 
@@ -131,6 +121,8 @@ echo "Done";
 
 function CreateXML($file, $bookid, $counter)
 {
+global $linefeed;
+global $space;
 
 	$xml = '<book xmlns="http://archive.org/scribe/xml">' . "\r\n";
 	$xml = $xml . '<bookData>' . "\r\n";
@@ -138,14 +130,15 @@ function CreateXML($file, $bookid, $counter)
 	$xml = $xml . '</bookData>' . "\r\n";
 	$xml = $xml . '<pageData>' . "\r\n";
 	
-	echo "Looking for file: " .  $file;
+	echo "Looking for file: " .  $file . $linefeed;
 	
 		if (file_exists($file)) {
 			$backupplan = false;
-			echo "$space$space$space$spaceFound Abbyy XML" . $linefeed;
-			
+			echo $space . $space . $space . $space . "Found Abbyy XML" . $linefeed;
+
 				try {
 				$abbyy = new SimpleXMLElement($file, NULL, TRUE);
+
 					$xmlpage = 1;
 					
 					foreach ($abbyy->page as $page) {
@@ -182,7 +175,7 @@ function CreateXML($file, $bookid, $counter)
 
 						$xmlpage = $xmlpage + 1;
 					}
-					echo "$space$space$space$spaceTotal XML Page elements processed for Abbyy: " . ($xmlpage - 1) . $linefeed;
+					echo $space . $space . $space . $space . "Total XML Page elements processed for Abbyy: " . ($xmlpage - 1) . $linefeed;
 				}	catch(Exception $e) {
 				
 					echo "Error Loading Abbyy";
@@ -194,7 +187,7 @@ function CreateXML($file, $bookid, $counter)
 			}
 			
 			if ($backupplan) {
-				echo "$space$space$space$spaceAbbyy xml not found or corrupt going to do it the hard way." . $linefeed;
+				echo $space . $space . $space . $space . "Abbyy xml not found or corrupt going to do it the hard way." . $linefeed;
 			
 				for ($i=1; $i<=$counter; $i++)
 					  {
@@ -234,43 +227,45 @@ function CreateXML($file, $bookid, $counter)
 
 function CreateFilesXML($bookid, $dir)
 {
+global $linefeed;
+global $space;
 
 	$xml = '<?xml version="1.0" encoding="UTF-8"?>' . "\r\n";
 	$xml = $xml . '<files>' . "\r\n";
 	
-		if (file_exists($dir . $bookid . '/' . $bookid . '_raw_jp2.zip')) {
+		if (file_exists($dir . '/' . $bookid . '/' . $bookid . '_raw_jp2.zip')) {
 			$xml = $xml . '<file name="' . $bookid . '_raw_jp2.zip" source="original">' . "\r\n";
 			$xml = $xml . '<format>Single Page Raw JP2 ZIP</format>' . "\r\n";
 			$xml = $xml . '</file>' . "\r\n";
 		}
 	
-		if (file_exists($dir . $bookid . '/' . $bookid . '_jp2.zip')) {
+		if (file_exists($dir . '/' . $bookid . '/' . $bookid . '_jp2.zip')) {
 			$xml = $xml . '<file name="' . $bookid . '_jp2.zip" source="derivative">' . "\r\n";
 			$xml = $xml . '<format>Single Page Processed JP2 ZIP</format>' . "\r\n";
 			$xml = $xml . '<original>' . $bookid . '_raw_jp2.zip</original>' . "\r\n";
 			$xml = $xml . '</file>' . "\r\n";
 		}
-		if (file_exists($dir . $bookid . '/' . $bookid . '_meta.xml')) {
+		if (file_exists($dir . '/' . $bookid . '/' . $bookid . '_meta.xml')) {
 			$xml = $xml . '<file name="' . $bookid . '_meta.xml" source="metadata">' . "\r\n";
 			$xml = $xml . '<format>Metadata</format>' . "\r\n";
 			$xml = $xml . '</file>' . "\r\n";
 		}
-		if (file_exists($dir . $bookid . '/' . $bookid . '_files.xml')) {
+		if (file_exists($dir . '/' . $bookid . '/' . $bookid . '_files.xml')) {
 			$xml = $xml . '<file name="' . $bookid . '_files.xml" source="metadata">' . "\r\n";
 			$xml = $xml . '<format>Metadata</format>' . "\r\n";
 			$xml = $xml . '</file>' . "\r\n";
 		}
-		if (file_exists($dir . $bookid . '/' . $bookid . '_cover_image.jpg')) {
+		if (file_exists($dir . '/' . $bookid . '/' . $bookid . '_cover_image.jpg')) {
 			$xml = $xml . '<file name="' . $bookid . '_cover_image.jpg" source="original">' . "\r\n";
 			$xml = $xml . '<format>Book Cover Image</format>' . "\r\n";
 			$xml = $xml . '</file>' . "\r\n";
 		}
-		if (file_exists($dir . $bookid . '/' . $bookid . '_abbyy.gz')) {
+		if (file_exists($dir . '/' . $bookid . '/' . $bookid . '_abbyy.gz')) {
 			$xml = $xml . '<file name="' . $bookid . '_abbyy.gz" source="original">' . "\r\n";
 			$xml = $xml . '<format>OCR</format>' . "\r\n";
 			$xml = $xml . '</file>' . "\r\n";
 		}
-		if (file_exists($dir . $bookid . '/' . $bookid . '_abbyy.xml')) {
+		if (file_exists($dir . '/' . $bookid . '/' . $bookid . '_abbyy.xml')) {
 			$xml = $xml . '<file name="' . $bookid . '_abbyy.xml" source="original">' . "\r\n";
 			$xml = $xml . '<format>OCR</format>' . "\r\n";
 			$xml = $xml . '</file>' . "\r\n";
@@ -281,55 +276,76 @@ function CreateFilesXML($bookid, $dir)
 	return $xml;
 }
 
-
 function CreateMetadataFilesXML($bookid, $SQLIPAddress, $UserName, $Password, $db)
 {
+global $linefeed;
+global $space;
 
-	$link = mysqli_connect($SQLIPAddress,$UserName,$Password) or 
-		die("Couldn't connect to SQL Server on $SQLIPAddress"); 
+	$link = new PDO("dblib:host=" . $SQLIPAddress . ";dbname=" . $db, $UserName, $Password) or 
+		die("Couldn't connect to SQL Server on $SQLIPAddress");
 
-	mssqli_select_db($db, $link) or 
-		die("Couldn't connect to database on $db"); 
-		
 	$query = "SELECT id, yb_id, title, year, creator, [description] as descrip, bookrecord, users_id, last_update, status FROM yb_metadata where yb_id = '" . $bookid . "'";
 
-	foreach ($mssqlconnection->query($query) as $row) { 
+	echo $query;
+
+	$row_count = 0;
+
+		foreach ($link->query($query) as $row) {
+			echo "<br />" . "I found a record" . "<br />";
+			$row_count = $row_count + 1;
 			
-		$xml = '<?xml version="1.0" encoding="UTF-8"?>' . "\r\n";
-		$xml = $xml . '<metadata>' . "\r\n";
-		$xml = $xml . '<identifier>' . trim($row['yb_id']) . '</identifier>' . "\r\n";
-		$xml = $xml . '<title>' . trim($row['title']) . '</title>' . "\r\n";
-		$xml = $xml . '<creator>' . trim($row['creator']) . '</creator>' . "\r\n";
-		$xml = $xml . '<description>' . trim(str_replace('<br>', '<br />', $row['descrip'])) . '</description>' . "\r\n";    
-		$xml = $xml . '<bookrecord>' . trim($row['bookrecord']) . '</bookrecord>' . "\r\n";   
-		$xml = $xml . '<year>' . trim($row['year']) . '</year>' . "\r\n";   	
-		$xml = $xml . '</metadata>' . "\r\n";
-		
-	} 
+			$xml = '<?xml version="1.0" encoding="UTF-8"?>' . "\r\n";
+			$xml = $xml . '<metadata>' . "\r\n";
+			$xml = $xml . '<identifier>' . trim($row['yb_id']) . '</identifier>' . "\r\n";
+			$xml = $xml . '<title>' . trim($row['title']) . '</title>' . "\r\n";
+			$xml = $xml . '<creator>' . trim($row['creator']) . '</creator>' . "\r\n";
+			$xml = $xml . '<description>' . trim(str_replace('<br>', '<br />', $row['descrip'])) . '</description>' . "\r\n";    
+			$xml = $xml . '<bookrecord>' . trim($row['bookrecord']) . '</bookrecord>' . "\r\n";   
+			$xml = $xml . '<year>' . trim($row['year']) . '</year>' . "\r\n";   	
+			$xml = $xml . '</metadata>' . "\r\n";
+			
+		} 
 	
-	mssql_close($link);		
+	if ($row_count == 0) {
+		echo "<br />";
+		echo "I don't have one: " . $row_count . "<br />";
+
+		$query = "INSERT into yb_metadata (yb_id, showcover, showyearbook) values ('" . $bookid . "', 0, 0)";
+		$link->query($query);
+
+			$xml = '<?xml version="1.0" encoding="UTF-8"?>' . "\r\n";
+			$xml = $xml . '<metadata>' . "\r\n";
+			$xml = $xml . '<identifier>' . trim($bookid) . '</identifier>' . "\r\n";
+			$xml = $xml . '<title></title>' . "\r\n";
+			$xml = $xml . '<creator></creator>' . "\r\n";
+			$xml = $xml . '<description></description>' . "\r\n";    
+			$xml = $xml . '<bookrecord></bookrecord>' . "\r\n";   
+			$xml = $xml . '<year></year>' . "\r\n";   	
+			$xml = $xml . '</metadata>' . "\r\n";
+		
+	}
 	
 	return $xml;
+	
 }
 
 function CreateImage($bookid, $dir, $tmpfolder)
 {
+global $linefeed;
+global $space;
+
 	try {
 //		echo $dir . $bookid . '/' . $bookid . '_0001.jp2' . $linefeed;
 //		echo $dir . $bookid . '/' . $bookid . '_cover_image.jpg'. $linefeed;
 
-		echo $tmpfolder . $bookid . '/' . $bookid . '_0001.jp2' . $linefeed;
-		echo $tmpfolder . $bookid . '/' . $bookid . '_cover_image.jpg'. $linefeed;
+		echo $dir . '/' . $bookid . '/' . $bookid . '_0001.jp2' . $linefeed;
+		echo $dir . '/' . $bookid . '/' . $bookid . '_cover_image.jpg'. $linefeed;
 		
-		echo "x";
 		$image = new Imagick();
-		echo "y";
-		//$image->readImage( $dir . '/' . $bookid . '/' . $bookid . '_0001.jp2' );
-		$image->readImage( $tmpfolder . $bookid . '/' . $bookid . '_0001.jp2' );		
+		$image->readImage( $dir . '/' . $bookid . '/' . $bookid . '_0001.jp2' );
 		$image->setImageFormat('jpg');
 		$image->adaptiveResizeImage(720,965);
-//		$image->writeImage($dir . '/' . $bookid . '/' . $bookid . '_cover_image.jpg');	
-		$image->writeImage($tmpfolder . $bookid . '/' . $bookid . '_cover_image.jpg');			
+		$image->writeImage($dir . '/' . $bookid . '/' . $bookid . '_cover_image.jpg');	
 		echo "Cover Created";
 		return;
 	}
@@ -340,19 +356,19 @@ function CreateImage($bookid, $dir, $tmpfolder)
 
 function CreateImageCMD($bookid, $dir, $tmpfolder)
 {
-	try {
-		//$input = $dir . '/' . $bookid . '/' . $bookid . '_0001.jp2';
-		//$output = $dir . '/' . $bookid . '/' . $bookid . '_cover_image.jpg';
+global $linefeed;
+global $space;
 
-		$input = $tmpfolder . $bookid . '/' . $bookid . '_0001.jp2';
-		$output = $tmpfolder . $bookid . '/' . $bookid . '_cover_image.jpg';
-		
+	try {
+		$input = $dir . '/' . $bookid . '/' . $bookid . '_0001.jp2';
+		$output = $dir . '/' . $bookid . '/' . $bookid . '_cover_image.jpg';
+
 		$thecmd = 'convert -resize 15% ' . $input . ' ' . $output;
 //		echo $thecmd;
 		exec($thecmd, $info);
 		
 		//echo $linefeed;
-		echo "$space$space$spaceCover Created" . $linefeed;
+		echo $space . $space . $space . "Cover Created" . $linefeed;
 		return;
 	}
 	catch(Exception $e) {
@@ -362,12 +378,14 @@ function CreateImageCMD($bookid, $dir, $tmpfolder)
 
 function CreateZip($bookid, $dir, $imageloc)
 {
+global $linefeed;
+global $space;
+
 	$zip = new ZipArchive();
 	
-	//$filename = $dir . '/' . $bookid . '/scandata.zip';
-	$filename = '/tmp' . '/' . $bookid . '/scandata.zip';
-
-	echo "$space$space$space$space$space$space" . $dir . $bookid . '/scandata.zip' . $linefeed;
+	$filename = $dir . '/' . $bookid . '/scandata.zip';
+	
+	echo $space . $space .$space .$space .$space .$space . $dir . '/' . $bookid . '/scandata.zip' . $linefeed;
 	
 		if ($zip->open($filename, ZIPARCHIVE::CREATE | ZIPARCHIVE::OVERWRITE)!==TRUE) {
 			exit("cannot open <$filename>\n");
@@ -378,10 +396,10 @@ function CreateZip($bookid, $dir, $imageloc)
 	
 	$zip->addFile($imageloc . "l.jpg", "l.jpg");
 	$zip->addFile($imageloc . "r.jpg", "r.jpg");
-	$zip->addFile($dir . $bookid . '/' . $bookid . '_cover_image.jpg', '' . $bookid . '_cover_image.jpg');
-	$zip->addFile($dir . $bookid . '/scandata.xml', 'scandata.xml');
+	$zip->addFile($dir . '/' . $bookid . '/' . $bookid . '_cover_image.jpg', '' . $bookid . '_cover_image.jpg');
+	$zip->addFile($dir . '/' . $bookid . '/' . 'scandata.xml', 'scandata.xml');
 
-	echo "$space$space$space$space$space$spaceNumber of Files added to zip: " . $zip->numFiles . $linefeed;
+	echo $space . $space .$space .$space .$space .$space . "Number of Files added to zip: " . $zip->numFiles . $linefeed;
 	//echo "status:" . $zip->status . $linefeed;
 	echo $linefeed;
 	$zip->close();
@@ -391,37 +409,28 @@ function CreateZip($bookid, $dir, $imageloc)
 
 function CreateAbbyygz($bookid, $dir, $tmpfolder)
 {
-	//$filename = $dir . '/' . $bookid . '_abbyy.gz' . $linefeed;
-	
-	$filename = $tmpfolder . $bookid . '/' . $bookid . '_abbyy.tar';
-	$filename2 = $tmpfolder . $bookid . '/' . $bookid . '_abbyy.gz';	
-	//echo $dir; = /mnt/yearbooks/
+global $linefeed;
+global $space;
+
+	$filename = $dir . '/' . $bookid . '/' . $bookid . '_abbyy.gz';	
+	$filename2 = $dir . '/' . $bookid . '/' . $bookid . '_abbyy.gz';	
+
 	// Creating the .gz with the double extension saves the extension inside of the .gz
 	
-	$abbyxmlfile = $dir . $bookid . '/' . $bookid . '_abbyy.xml';
+	$abbyxmlfile = $dir . '/' . $bookid . '/' . $bookid . '_abbyy.xml';
 	
-	echo "$space$space$space$spacegz Location " . $filename . $linefeed;
-	echo "$space$space$space$spaceXML Location " . $abbyxmlfile . $linefeed;
+	echo $space . $space .$space .$space . "gz Location " . $filename . $linefeed;
+	echo $space . $space .$space .$space . "_abbyy.gz location " . $filename2 . $linefeed;
+	echo $space . $space .$space .$space . "XML Location " . $abbyxmlfile . $linefeed;
 
 	if (file_exists($abbyxmlfile)) {
 	
 		try
 		{
-		
-			//$thecmd = 'tar -cvf ' . $filename . ' -C ' . $dir . $bookid . '/ ' . $bookid . '_abbyy.xml';
-			//echo $thecmd;
-			//exec($thecmd, $info);
 			
-			//$thecmd = 'gzip ' . $filename . ' ' . $filename2;
-			//echo $thecmd;
-			//exec($thecmd, $info);
-			
-			//echo $dir . $bookid . '/' . $bookid . '_abbyy.xml';
-			
-			$thecmd = 'gzip -c ' . $dir . $bookid . '/' . $bookid . '_abbyy.xml' . ' > ' . $filename2;
+			$thecmd = 'gzip -c ' . $dir . '/' . $bookid . '/' . $bookid . '_abbyy.xml' . ' > ' . $filename2;
 			echo $thecmd;
 			exec($thecmd, $info);
-			
 			
 		
 		} 
@@ -431,7 +440,7 @@ function CreateAbbyygz($bookid, $dir, $tmpfolder)
 		}
 		
 	} else {
-		echo "$space$space$space$spaceNo _Abbyy.XML found" . $linefeed;
+		echo $space . $space .$space .$space . "No _Abbyy.XML found" . $linefeed;
 	}
 	
 	return;
